Add a variable to count the comparisons
To know the comparisons made in qsort. A global variable "cmp_cnt" is
added and will increase once the comparison function is called.

+++ sort.h
@@ -14,6 +14,6 @@ void sort_main(void *sort_buffer,
                size_t size,
                size_t es,
                cmp_t cmp,
-               int sort_type);
+               size_t sort_type);
 
 #endif
